collection-service/functions/remove_card/requirements.txt
aws_xray_sdk
boto3
python-jose